\name{bees}

\alias{bees}

\docType{data}

\title{bees}

\description{
  Data from ...
}

\usage{data(bees)}

\format{
  A data frame with 66 observations on the following 5 variables.
  \describe{
    \item{\code{mixture}}{}
    \item{\code{treat}}{}
    \item{\code{rep}}{}
    \item{\code{dead0h}}{}
    \item{\code{dead48h}}{}
  }
}

\details{
  
}

\source{
  
}

\references{
  }

\examples{

}
\keyword{datasets}
