\name{Daphnia}

\alias{Daphnia}

\docType{data}

\title{Daphnia}

\description{
  Data from ...
}

\usage{data(Daphnia)}

\format{
  A data frame with 140 observations on the following 6 variables.
  \describe{
    \item{\code{dose.a}}{}
    \item{\code{dose.p}}{}
    \item{\code{dose}}{}
    \item{\code{mix.frac}}{}
    \item{\code{total}}{}
    \item{\code{immob48}}{}
  }
}

\details{
  
}

\source{
  
}

\references{
  }

\examples{

}
\keyword{datasets}
