\name{Eryngium.sparganophyllum}
\alias{Eryngium.sparganophyllum}
\docType{data}
\title{}
\description{}

\usage{data("Eryngium.sparganophyllum")}

\format{
  A data frame with 583 observations on the following 5 variables.
  \describe{
    \item{\code{Day}}{a numeric vector}
    \item{\code{Treat}}{a factor with 15 levels}
    \item{\code{Type}}{a factor with two levels}
    \item{\code{Rep}}{a numeric vector}
    \item{\code{Germ}}{a numeric vector}
  }
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}

\keyword{datasets}
