\name{red.fescue}

\alias{red.fescue}

\docType{data}

\title{Red fescue}

\description{
  
}

\usage{data(red.fescue)}

\format{
  A data frame with 50 observations on the following 5 variables.
  \describe{
    \item{\code{Species}}{a categorial vector}
    \item{\code{Dose}}{a numeric vector}
    \item{\code{Time}}{a categorical vector}
    \item{\code{Biomass}}{a numeric vector}
    \item{\code{Curve}}{a categorial vector}
  }
}

\details{
  
}

\source{
  
}

%\references{}

\examples{
}
\keyword{datasets}
