\name{ryegrass2}

\alias{ryegrass2}

\docType{data}

\title{Ryegrass}

\description{ryegrass2 data.}

\usage{data(ryegrass2)}

\format{
  A data frame with 27 observations on the following 4 variables.
  \describe{
    \item{\code{Dose}}{a numeric vector}
    \item{\code{wheat}}{a numeric vector}
    \item{\code{weed}}{a numeric vector}
    \item{\code{curve}}{a categorial vector}
  }
}

\keyword{datasets}
