\name{red.fescue}

\alias{red.fescue}

\docType{data}

\title{Red fescue...}

\description{
  
}

\usage{data(red.fescue)}

\format{
  A data frame with 26 observations on the following 2 variables.
  \describe{
    \item{\code{Dose}}{a numeric vector}
    \item{\code{Biomass}}{a numeric vector}
  }
}

\details{
  
}

\source{
  
}

%\references{}

\examples{
}
\keyword{datasets}
